refactor the whole hypervisor and userspace code
this commit do lots of changes for almost all the code, the
purpose of this refactor is to better support for
porting Minos hypervisor to more platform and board.
after this refactor, the hypervisor has a more powerful
framework to support more virtual device. this commit
only test on arm-fvp platform.

Signed-off-by: Min le <lemin9538@gmail.com>