add ramdisk code to support simple ramdisk feature
with this feature, the uboot will load only 3 files
- minos.bin
- minos.dtb
- minos.img (ramdisk image)

the minos.img will contain all the images of the VM
and all the dtb file for each VM if it has. in each
vm's dtb entry will add two member to point which file
the vm will use:
- image-file = "vm0-boot.img"
- dtb-file = "vm0-dtb.img"

the minos.dtb will have two entry in chosen to record the
information of the ramdisk.
- minos,initrd-start (start address in memory)
- minos,initrd-end   (end address in memory)

Signed-off-by: Min Le <lemin9538@gmail.com>